requirements.txt
# Data Validation Solution - Python Dependencies

# Core data processing
pandas>=2.0.0
openpyxl>=3.1.0

# Configuration
PyYAML>=6.0
python-dotenv>=1.0.0

# Database connectors
pyodbc>=5.0.0
cx-Oracle>=8.3.0
nzpy>=1.1.0
snowflake-connector-python>=3.0.0

# Utilities
colorama>=0.4.6

# Optional: For CSV SQL queries
pandasql>=0.7.3
